Extract query history scrubber tests
